Update ProductInfo page.
Show product image.
Show product price.
Show product description.
Enable to add to the cart.
